Add files via upload
It is a number guessing game, by using Math.random() method it generates one number we have to guess the number in K trails. If the number we guessed is greater it prints number is greater. If it is less it prints less than the number. If we guess the number it prints Congratulations.